Improved root.tsx
