--- app/components/Error/index.ts
@@ -1,1 +1,0 @@
-export { Error } from "./Error.tsx"
