--- app/components/System/index.ts
@@ -1,1 +1,0 @@
-export { System } from "./System.tsx"
