+++ app/locales/en.json
@@ -1,1 +1,7 @@
-{}
+{
+  "Page not found": "Page not found",
+  "Something went wrong": "Something went wrong",
+  "Take me back to home page": "Take me back to home page",
+  "error404": "You may have mistyped the address, or the page has been moved to another URL.",
+  "error500": "Something went wrong on our side, please try again later or visit the home page."
+}
