+++ app/package.json
@@ -7,7 +7,9 @@
     "#*": "./*"
   },
   "scripts": {
-    "generate": "wrangler types"
+    "generate": "pnpm generate:cf && pnpm generate:rr",
+    "generate:cf": "wrangler types",
+    "generate:rr": "react-router typegen"
   },
   "dependencies": {
     "@mantine/core": "8.3.5",
@@ -21,7 +23,8 @@
     "react-schemaorg": "2.0.0",
     "remix-utils": "9.0.0",
     "tiny-invariant": "^1.3.3",
-    "ts-extras": "^0.14.0"
+    "ts-extras": "^0.14.0",
+    "xml-js": "1.6.11"
   },
   "devDependencies": {
     "@cloudflare/vite-plugin": "1.13.13",
