+++ app/routes/root.module.css
@@ -1,9 +1,9 @@
-.root {
+.errorRoot {
   padding-top: rem(80px);
   padding-bottom: rem(80px);
 }
 
-.label {
+.errorLabel {
   text-align: center;
   font-weight: 900;
   line-height: 1;
@@ -16,17 +16,15 @@
   }
 }
 
-.description {
+.errorDescription {
   max-width: rem(500px);
   margin: auto;
   margin-top: var(--mantine-spacing-xl);
   margin-bottom: calc(1.5 * var(--mantine-spacing-xl));
 }
 
-.title {
-  font-family:
-    Greycliff CF,
-    var(--mantine-font-family);
+.errorTitle {
+  font-family: var(--mantine-font-family);
   text-align: center;
   font-weight: 900;
   font-size: rem(32px);
