+++ package.json
@@ -40,6 +40,7 @@
     "tempy": "3.1.0",
     "type-fest": "^4.41.0",
     "typescript": "5.9.2",
+    "vite": "7.1.10",
     "vitest": "3.2.4"
   },
   "packageManager": "pnpm@10.11.0+sha512.6540583f41cc5f628eb3d9773ecee802f4f9ef9923cc45b69890fb47991d4b092964694ec3a4f738a420c918a333062c8b925d312f42e4f0c263eb603551f977"
